requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-filter==24.3
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
drf-yasg==1.21.8